Removed broken functionality, Modified Proof Of Concept.
Tasker was changed to `shell` which will run a single shell command in another thread and return the result if the client remains connected.

Accepts a string now rather than an array (because shell=True for the subprocess)

Removed functionality relating to creating named shell tasks because the change to seperate files for actions broke the list of tasks in memory (???) Idk. Don't need that functionality anyway, though it is neat.

Changed some print statements to logger statements instead.

Removed actions.py in its entirety and moved the relevant functionality to "DoAction" in receiver.py.

The Proof of concept script was modified to reflect the loss of task functionality. This makes it look much more streamlined, :thought_balloon: